requirements.txt
discord.py==1.2.5
python-dotenv==0.10.3
pymongo==3.10.1
dnspython==1.16.0
PyDictionary==2.0.1
urbandictionary==1.1